// generated/src/app/api-models/al-letter-template.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AlLetterTemplate {
  seqLetterTempId: number;
  letterType: string;
  letterTemplateId: string;
  description: string;
  letterGroupFlag: string;
  overrideDistributionMethod: string;
  primaryDistributionSource: string;
  viewHistoryFlag: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}